// types/widgets/slider.d.ts
export type EventHandler = (self: AgsSlider, event: Gdk.Event) => void | unknown;
declare const POSITION: {
    readonly left: Gtk.PositionType.LEFT;
    readonly right: Gtk.PositionType.RIGHT;
    readonly top: Gtk.PositionType.TOP;
    readonly bottom: Gtk.PositionType.BOTTOM;
};
export type Position = keyof typeof POSITION;
export type Mark = [number, string?, Position?] | number;
export type SliderProps = BaseProps<AgsSlider, Gtk.Scale.ConstructorProperties & {
    on_change?: EventHandler;
    value?: number;
    min?: number;
    max?: number;
    step?: number;
    marks?: Mark[];
}>;
// ...
export default class AgsSlider extends AgsSlider_base {
    constructor({ value, min, max, step, marks, ...rest }?: SliderProps);
    get on_change(): EventHandler;
    set on_change(callback: EventHandler);
    get value(): number;
    set value(value: number);
    get min(): number;
    set min(min: number);
    get max(): number;
    set max(max: number);
    get step(): number;
    set step(step: number);
    set marks(marks: Mark[]);
    get dragging(): boolean;
    set dragging(dragging: boolean);
    get vertical(): boolean;
    set vertical(vertical: boolean);
    vfunc_button_release_event(event: Gdk.EventButton): boolean;
    vfunc_button_press_event(event: Gdk.EventButton): boolean;
    vfunc_key_press_event(event: Gdk.EventKey): boolean;
    vfunc_key_release_event(event: Gdk.EventKey): boolean;
    vfunc_scroll_event(event: Gdk.EventScroll): boolean;
}