// types/widgets/stack.d.ts
declare const TRANSITION: {
    readonly none: Gtk.StackTransitionType.NONE;
    readonly crossfade: Gtk.StackTransitionType.CROSSFADE;
    readonly slide_right: Gtk.StackTransitionType.SLIDE_RIGHT;
    readonly slide_left: Gtk.StackTransitionType.SLIDE_LEFT;
    readonly slide_up: Gtk.StackTransitionType.SLIDE_UP;
    readonly slide_down: Gtk.StackTransitionType.SLIDE_DOWN;
    readonly slide_left_right: Gtk.StackTransitionType.SLIDE_LEFT_RIGHT;
    readonly slide_up_down: Gtk.StackTransitionType.SLIDE_UP_DOWN;
    readonly over_up: Gtk.StackTransitionType.OVER_UP;
    readonly over_down: Gtk.StackTransitionType.OVER_DOWN;
    readonly over_left: Gtk.StackTransitionType.OVER_LEFT;
    readonly over_right: Gtk.StackTransitionType.OVER_RIGHT;
    readonly under_up: Gtk.StackTransitionType.UNDER_UP;
    readonly under_down: Gtk.StackTransitionType.UNDER_DOWN;
    readonly under_left: Gtk.StackTransitionType.UNDER_LEFT;
    readonly under_right: Gtk.StackTransitionType.UNDER_RIGHT;
    readonly over_up_down: Gtk.StackTransitionType.OVER_UP_DOWN;
    readonly over_down_up: Gtk.StackTransitionType.OVER_DOWN_UP;
    readonly over_left_right: Gtk.StackTransitionType.OVER_LEFT_RIGHT;
    readonly over_right_left: Gtk.StackTransitionType.OVER_RIGHT_LEFT;
};
export type Transition = keyof typeof TRANSITION;
export type StackProps = BaseProps<AgsStack, Gtk.Stack.ConstructorProperties & {
    shown?: string;
    items?: [string, Gtk.Widget][];
    transition?: Transition;
}>;
// ...
export default class AgsStack extends AgsStack_base {
    constructor(props?: StackProps);
    add_named(child: Gtk.Widget, name: string): void;
    get items(): [string, Gtk.Widget][];
    set items(items: [string, Gtk.Widget][]);
    get transition(): Transition;
    set transition(transition: Transition);
    get shown(): string | null;
    set shown(name: string | null);
}